+++ CHANGES.rst
@@ -8,6 +8,8 @@ Enhancements and Fixes
 
 - Fix job result handling to prioritize standard URL structure [#684]
 
+- Fix DALOverflowWarning error message to only indicate the cause as limits from user or server. [#689]
+
 Deprecations and Removals
 -------------------------